Adds a new RecipeReader class as the basis of all Recipe Parsing classes and refactors RecipeParser.
