+++ tests/constants.py
@@ -1,0 +1,14 @@
+"""
+:Description: Contains general test constants
+"""
+
+from __future__ import annotations
+
+from typing import Final
+
+# Long multi-line description string found in the `simple-recipe.yaml` test file
+SIMPLE_DESCRIPTION: Final[str] = (
+    "This is a PEP '561 type stub package for the toml package.\n"
+    "It can be used by type-checking tools like mypy, pyright,\n"
+    "pytype, PyCharm, etc. to check code that uses toml."
+)
